+++ epip/decision/exceptions.py
@@ -1,0 +1,13 @@
+"""EPIP-012 exceptions."""
+
+
+class DecisionError(Exception):
+    pass
+
+
+class InvalidDecisionInputError(DecisionError):
+    pass
+
+
+class DecisionVersionError(DecisionError):
+    pass